refactor: rename workspace preferences folder to workspace management
